# FE/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sports Center Management API",
    "version": "1.0.0",
    "description": "Backend API for Sports Center Management System – Flow 1 (User & Membership), Flow 2 (Class & Schedule), Flow 3 (Payment & Report)"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    },
    {
      "url": "https://sports-center-management-system.onrender.com/api/v1",
      "description": "Production Server (Render)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request (validation or malformed body)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Validation failed",
                "errors": [
                  {
                    "field": "email",
                    "message": "Invalid email address"
                  }
                ]
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, invalid or expired token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Unauthorized: invalid or expired token"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient role permissions",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Forbidden: insufficient permissions"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Record not found"
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "Duplicate value or business conflict",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Duplicate value for: email"
              }
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Internal server error"
              }
            }
          }
        }
      },
...
      "RoomListOk": {
        "description": "Paginated list of rooms (compact example)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": true,
                "message": "Rooms retrieved successfully",
                "data": [
                  {
                    "id": "c3e1ef3e-0000-4000-8000-000000000101",
                    "name": "Yoga Room A",
                    "capacity": 20,
                    "location": "Floor 1",
                    "areaType": "INDOOR",
                    "isActive": true
                  }
                ],
                "pagination": {
                  "page": 1,
                  "limit": 10,
                  "total": 2,
                  "totalPages": 1
                }
              }
            }
          }
        }
      },
      "RoomCreated": {
        "description": "Room created",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": true,
                "message": "Room created successfully",
                "data": {
                  "id": "c3e1ef3e-0000-4000-8000-000000000103",
                  "name": "Boxing Room",
                  "capacity": 12,
                  "areaType": "INDOOR",
                  "isActive": true
                }
              }
            }
          }
        }
      },
      "RoomOk": {
        "description": "Single room",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": true,
                "message": "Room retrieved successfully",
                "data": {
                  "id": "c3e1ef3e-0000-4000-8000-000000000101",
                  "name": "Yoga Room A",
                  "capacity": 20,
                  "location": "Floor 1",
                  "areaType": "INDOOR",
                  "isActive": true
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/rooms": {
      "get": {
        "summary": "Get list of rooms",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by room name"
          },
          {
            "in": "query",
            "name": "areaType",
            "schema": {
              "type": "string",
              "enum": [
                "POOL",
                "INDOOR",
                "OUTDOOR"
              ]
            },
            "description": "Filter by area type (POOL | INDOOR | OUTDOOR)"
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/RoomListOk"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "summary": "Create a new room",
        "tags": [
          "Rooms"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "capacity",
                  "areaType"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Yoga Room A"
                  },
                  "capacity": {
                    "type": "integer",
                    "example": 20
                  },
                  "location": {
                    "type": "string",
                    "example": "Floor 1"
                  },
                  "areaType": {
                    "type": "string",
                    "enum": [
                      "POOL",
                      "INDOOR",
                      "OUTDOOR"
                    ],
                    "example": "INDOOR"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/RoomCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/rooms/{id}": {
      "get": {
        "summary": "View room details",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/RoomOk"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "patch": {
        "summary": "Update room info",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "capacity": {
                    "type": "integer"
                  },
                  "location": {
                    "type": "string"
                  },
                  "areaType": {
                    "type": "string",
                    "enum": [
                      "POOL",
                      "INDOOR",
                      "OUTDOOR"
                    ],
                    "description": "New area type. Upcoming schedules must use a matching Class."
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/RoomOk"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "summary": "Deactivate room (soft delete)",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/RoomOk"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/rooms/{roomId}/transfer-schedules/preview": {
      "post": {
        "summary": "Preview bulk transfer of upcoming schedules to another room (no DB changes)",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "roomId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Source room (damaged room)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetRoomId"
                ],
                "properties": {
                  "targetRoomId": {
                    "type": "string",
                    "description": "Target room ID"
                  },
                  "from": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional lower bound (default now)"
                  },
                  "to": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional upper bound"
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "Transfer reason, e.g. Room damaged"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preview with totalSchedules/validSchedules/invalidSchedules/canTransfer/conflicts"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/rooms/{roomId}/transfer-schedules": {
      "post": {
        "summary": "Bulk transfer upcoming SCHEDULED schedules to another room (atomic, all-or-nothing)",
        "description": "Validates areaType/capacity/room+coach conflicts per schedule, re-checks inside a Prisma transaction, then notifies enrolled members with SCHEDULE_ROOM_CHANGED.",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "roomId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Source room (damaged room)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetRoomId"
                ],
                "properties": {
                  "targetRoomId": {
                    "type": "string",
                    "description": "Target room ID"
                  },
                  "from": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional lower bound (default now)"
                  },
                  "to": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional upper bound"
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "Transfer reason, e.g. Room damaged"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transferred with sourceRoom/targetRoom/transferredCount/schedules"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}